fix: 댓글 작성 완료 시 alert 호출 및 CommentContainer 분리
