+++ src/app/studyDetail/studyContent/ask/comments/CommentList.tsx
@@ -1,8 +1,6 @@
 import { Comment } from '@src/api/types';
-import InputBase from '@src/app/shared/components/input/InputBase';
-import usePostStudyComment from '@src/hooks/remotes/comment/usePostStudyComment';
-import React, { useMemo, useState } from 'react';
-import CommentItem from './CommentItem';
+import React, { useMemo } from 'react';
+import CommentItemContainer from './CommentItemContainer';
 
 interface CommentListProps {
 	comments: Comment[];
@@ -10,7 +8,7 @@ interface CommentListProps {
 	studyId: string;
 }
 
-interface ICommentWithNestedComments extends Comment {
+export interface ICommentWithNestedComments extends Comment {
 	nestedComments: Comment[];
 }
 
@@ -33,46 +31,9 @@ const CommentList = ({ comments, hostId, studyId }: CommentListProps): JSX.Eleme
 
 	return (
 		<div>
-			{commentsWithNestedComments.map((item) => {
-				const postComment = usePostStudyComment();
-
-				const [show, setShow] = useState<boolean>(false);
-				const [showCommentInput, setShowCommentInput] = useState<boolean>(false);
-				const [content, setContent] = useState<string>('');
-				const onSubmit = () => {
-					postComment.mutate({
-						id: studyId,
-						content,
-						replyTo: item.id,
-					});
-					setContent('');
-				};
-
-				return (
-					<div key={item.id} className="comment-list-item-container">
-						<CommentItem comment={item} hostId={hostId} setShowCommentInput={setShowCommentInput} />
-						{item.nestedComments?.map((nestedItem, nestedIndex) => {
-							if (!show && nestedIndex > 0) {
-								return <div />;
-							}
-							return <CommentItem key={nestedItem.id} comment={nestedItem} hostId={hostId} />;
-						})}
-						{item.nestedComments?.length > 1 && !show && (
-							<button type="button" className="comment-list-item-more-btn" onClick={() => setShow(!show)}>
-								더보기 ({item.nestedComments?.length - 1})
-							</button>
-						)}
-						{showCommentInput && (
-							<InputBase
-								placeholder="궁금한 점들을 물어보세요!"
-								content={content}
-								setContent={setContent}
-								onSubmit={onSubmit}
-							/>
-						)}
-					</div>
-				);
-			})}
+			{commentsWithNestedComments.map((item) => (
+				<CommentItemContainer key={item.id} studyId={studyId} hostId={hostId} item={item} />
+			))}
 		</div>
 	);
 };
