+++ src/hooks/remotes/comment/usePostStudyComment.ts
@@ -10,6 +10,7 @@ export default () => {
 
 	return useMutation(mutation, {
 		onSuccess: (response: any) => {
+			window.alert('댓글 작성에 성공하였습니다.');
 			console.log(response);
 		},
 		onError: (e: Error) => {
